feat: `GoToPageLink` 컴포넌트 구현 및 교체
